CMRNLP-91: Refactoring link generation
